feat: major api refactor, add typescript
BREAKING CHANGE: counsel has been completely rewritten. the intent has
not changed, but the implementation and interfaces as much different.
there is no migration guide. please take a peek at the new documentation
to decide if you want to migrate or not! the new interfaces are much
simpler, generally pure, and functional. i hope you enjoy it!